basic reply end

+++ What_Do_I_Sell_Today/src/main/webapp/WEB-INF/views/service/comment.jsp
@@ -1,5 +1,6 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 
 <script>
 	var boardNum = '${board.boardNum}'; //게시글 번호
@@ -36,9 +37,11 @@
 					$.each(result, function(key, value) {
 						a += '<div class="commentArea" style="border-bottom:1px solid darkgray; margin-bottom: 15px;">';
 						a += '<div class="commentInfo' + value.replyNum + '">' + '댓글번호 : ' + value.replyNum + ' / 작성자 : ' + value.usersID;
-						a += '<a onclick="commentUpdate(' + value.replyNum + ',\'' + value.contents + '\');"> 수정 </a>';
-						a += '<a onclick="commentDelete(' + value.replyNum + ');"> 삭제 </a> </div>';
-						a += '<div class="commentContent' + value.replyNum + '"> <p> 내용 : ' + value.contents + '</p>';
+						if('${userid}' == value.usersID) {
+							a += '<a onclick="commentUpdate(' + value.replyNum + ',\'' + value.contents + '\');"> 수정 </a>';
+							a += '<a onclick="commentDelete(' + value.replyNum + ');"> 삭제 </a>';
+						}
+						a += '</div> <div class="commentContent' + value.replyNum + '"> <p> 내용 : ' + value.contents + '</p>';
 						a += '</div></div>';
 					});
 					$(".commentList").html(a);
@@ -84,20 +87,23 @@
 
 	//댓글 삭제 
 	function commentDelete(replyNum) {
-		$.ajax({
-			type : 'delete',
-			url : '/deleteReply/' + replyNum,
-			headers : {
-				"Content-Type" : "application/json",
-				"X-HTTP-Method-Override" : "DELETE"
-			},
-			dataType : 'text',
-			success : function(result) {
-				if (result == 'Success') {
-					commentList();
+		var execute = confirm("정말 삭제하시겠습니까?");
+		if(execute) {
+			$.ajax({
+				type : 'delete',
+				url : '/deleteReply/' + replyNum,
+				headers : {
+					"Content-Type" : "application/json",
+					"X-HTTP-Method-Override" : "DELETE"
+				},
+				dataType : 'text',
+				success : function(result) {
+					if (result == 'Success') {
+						commentList();
+					}
 				}
-			}
-		});
+			});
+		}
 	}
 
 	$(document).ready(function() {
